clean folder structure
