+++ .gitignore
@@ -3,7 +3,6 @@ venv/
 requirements.txt
 *.pyc
 tests/
-models/
 # Python virtual environment
 
 # Byte-compiled / optimized / DLL files
