+++ app/models/rag.py
@@ -1,0 +1,19 @@
+import dspy
+
+class RAGResponse(dspy.Signature):
+    context: str = dspy.InputField()
+    prompt: str = dspy.InputField()
+    response: str = dspy.OutputField()
+
+class RAG(dspy.Module):
+    """Model to generate responses based on the retrieved context"""
+    def __init__(self, config: dict):
+        super().__init__()
+        self.model = config["model"]
+        self.temperature = config["temperature"]
+        self.max_tokens = config["max_tokens"]
+        RAGResponse.__doc__ = config["instruction"]
+        self.response = dspy.Predict(RAGResponse, temperature=self.temperature, max_tokens=self.max_tokens)
+    
+    async def forward(self, context: str = None, prompt: str = None) -> str:
+        response = await self.response.acall(context=context, prompt=prompt)
