--- modules/__init__.py
@@ -1,3 +1,0 @@
-from modules.text_processing.embedders import *
-from modules.text_processing.rag_engine import *
-from modules.llm_modules import *
