--- services/__init__.py
@@ -1,3 +1,0 @@
-# services/__init__.py
-from services.response_handlers import ResponseHandler
-
